+++ web/index.php
@@ -2,12 +2,25 @@
 declare(strict_types=1);
 
 use Ridibooks\Cms\Thrift\ThriftService;
+use Symfony\Component\HttpFoundation\Request;
 
 require_once __DIR__ . "/../vendor/autoload.php";
 
 if (is_readable(__DIR__ . '/../.env')) {
     $dotenv = new Dotenv\Dotenv(__DIR__, '/../.env');
     $dotenv->overload();
+}
+
+// If hostname has a form of dev domain, set test id.
+$request = Request::createFromGlobals();
+if (!empty($_ENV['TEST_AUTH_DISABLE'])) {
+    $hostname = $request->getHost();
+    $pattern = '/^admin\.(\w+)(\.platform)?\.dev\.ridi\.io$/';
+
+    // 'admin.{test_id}.dev.io' or 'admin.{test_id}.platform.dev.io'
+    if (preg_match($pattern, $hostname, $matches)) {
+        $_ENV['TEST_ID'] = $matches[1];
+    }
 }
 
 $cms_rpc_url = $_ENV['CMS_RPC_URL'] ?? '';
@@ -19,4 +32,4 @@ $config = require __DIR__ . '/../config/config.php';
 $app = require __DIR__ . '/../src/app.php';
 require __DIR__ . '/../src/controllers.php';
 
-$app->run();
+$app->run($request);